Downgrade cosmwasm-std to 0.16
Terra currently does not support the 1.0.0 beta, so we must dongrade to
0.16 and 0.8 for cosmwasm-storage-plus.

This has the downside of requiring us to change `mock_dependencies`, and
most notable the `query_opponent_matches` query must be changed to use
`ends_with` on the `Vec<u8>` byte string. This is insecure, however it
is the only option in cosmwasm 0.16.